+++ dev/dev-khulnasoft-server.sh
@@ -7,6 +7,6 @@ set -ex
 # that this image is not exactly identical to the published sourcegraph/server
 # images, as those include Khulnasoft Enterprise features.
 time cmd/server/pre-build.sh
-IMAGE=sourcegraph/server:0.0.0-DEVELOPMENT VERSION=0.0.0-DEVELOPMENT time cmd/server/build.sh
+IMAGE=khulnasoft/server:0.0.0-DEVELOPMENT VERSION=0.0.0-DEVELOPMENT time cmd/server/build.sh
 
-IMAGE=sourcegraph/server:0.0.0-DEVELOPMENT dev/run-server-image.sh
+IMAGE=khulnasoft/server:0.0.0-DEVELOPMENT dev/run-server-image.sh