+++ dev/src-expose/docker-publish.sh
@@ -12,10 +12,10 @@ export VERSION=$COMMIT_SHA
 export DATE
 DATE=$(date -u +"%Y-%m-%dT%H:%M:%SZ")
 
-export IMAGE=sourcegraph/src-expose
+export IMAGE=khulnasoft/src-expose
 
 . ./build.sh
 
-docker image push sourcegraph/src-expose:latest
+docker image push khulnasoft/src-expose:latest
 docker image tag sourcegraph/src-expose:latest sourcegraph/src-expose:"$VERSION"
-docker image push sourcegraph/src-expose:"$VERSION"
+docker image push khulnasoft/src-expose:"$VERSION"